bump lodash version
current lodash version has an identified vulnerability:
https://nodesecurity.io/advisories/577
this updates to latest version (that is patched for that vuln), and
updates usage on this lib,
luckily all breaking changes are just functions that got renamed.

they only tricky change was on test code, `_.cloneDeep` became smarter
and now produces that clone with same type (so `instanceof Resource` is
true), so I moved to use a dumber way of cloning for that test (JSON
parse+stringify).